+++ cached_path/schemes/beaker_v1.py
@@ -2,20 +2,21 @@ import io
 from pathlib import Path
 from typing import Optional
 
+from beaker import ChecksumFailedError  # type: ignore
+from beaker import DatasetNotFound  # type: ignore
+from beaker import DatasetReadError  # type: ignore
 from beaker import Beaker
-from beaker.exceptions import BeakerChecksumFailedError  # type: ignore
-from beaker.exceptions import BeakerDatasetNotFound  # type: ignore
 
 from .scheme_client import SchemeClient
 
 
 class BeakerClient(SchemeClient):
     scheme = ("beaker",)
-    recoverable_errors = SchemeClient.recoverable_errors + (BeakerChecksumFailedError,)
+    recoverable_errors = SchemeClient.recoverable_errors + (DatasetReadError, ChecksumFailedError)
 
     def __init__(self, resource: str) -> None:
         super().__init__(resource)
-
+        self.beaker = Beaker.from_env()
         # Beaker resources should be in the form "{user}/{dataset_name}/{path}/{to}/{file}"
         path = Path(resource.split("://")[1])
         if len(path.parts) < 2:
@@ -25,16 +26,15 @@ class BeakerClient(SchemeClient):
                 "or beaker://{dataset_id}/{path_to_file}."
             )
 
-        with Beaker.from_env() as beaker:  # type: ignore
-            try:
-                user, dataset_name, *filepath_parts = path.parts
-                self.dataset = beaker.dataset.get(f"{user}/{dataset_name}")
-            except BeakerDatasetNotFound:
-                dataset_id, *filepath_parts = path.parts
-                self.dataset = beaker.dataset.get(dataset_id)
+        try:
+            user, dataset_name, *filepath_parts = path.parts
+            self.dataset = self.beaker.dataset.get(f"{user}/{dataset_name}")
+        except DatasetNotFound:
+            dataset_id, *filepath_parts = path.parts
+            self.dataset = self.beaker.dataset.get(dataset_id)
 
-            self.filepath = "/".join(filepath_parts)
-            self.file_info = beaker.dataset.get_file_info(self.dataset, self.filepath)
+        self.filepath = "/".join(filepath_parts)
+        self.file_info = self.beaker.dataset.file_info(self.dataset, self.filepath)  # type: ignore
 
     def get_etag(self) -> Optional[str]:
         return None if self.file_info.digest is None else str(self.file_info.digest)
@@ -43,7 +43,6 @@ class BeakerClient(SchemeClient):
         return self.file_info.size
 
     def get_resource(self, temp_file: io.BufferedWriter) -> None:
-        with Beaker.from_env() as beaker:  # type: ignore
-            for chunk in beaker.dataset.stream_file(self.dataset, self.filepath):
-                if chunk:
-                    temp_file.write(chunk)
+        for chunk in self.beaker.dataset.stream_file(self.dataset, self.filepath, quiet=True):  # type: ignore
+            if chunk:
+                temp_file.write(chunk)