+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added support for beaker-py v2 for `beaker://` requests.
+
 ## [v1.7.4](https://github.com/allenai/cached_path/releases/tag/v1.7.4) - 2025-09-29
 
 - Added support for v1.0 of HuggingFace-Hub.